The code is now working
The code is fully functional but the output of area is always a float. I
am not sure the best way to change this, or if it is even necessary to
change it. I did modify the beginning of the code to make sure it
wouldnt error because of a string being inputted.